+++ src/entities/drawing.py
@@ -1,31 +1,33 @@
+import json
+
 class Drawing:
     """
         A drawing or a template modifiable by users
-        Basically a log for reproducing each change in order happened to the Canvas
+        Basically a log for reproducing each change in exact order as happened to the Canvas
     """
 
-    def __init__(self, name: str, width=800, height=600, content = None):
-        "creates a Drawing"
+    def __init__(self, name: str, width = 800, height = 600, id: int = None, content: list = None):
+        "creates a Drawing, width and height will be adjustable later prolly"
 
-        self._name = name
-        self._content = content
+        self.name = name
+        self.id = id
+        self.width = width
+        self.height = height
+        self._content = content if content else []
 
-    def add_rect(self):
-        "adds a rectangle"
-        pass
+    def add(self, cmd, *args, **kwargs):
+        "buffers used commands in a reproducible way, probably in JSON format in backend"
 
-    def add_oval(self):
-        "adds an oval"
-        pass
+        self._content.append([cmd, args, kwargs])
 
-    def add_line(self):
-        "adds a line"
-        pass
+    def reproduce(self):
+        "this could probably be a simple getter..."
 
-    def add_freeline(self):
-        "adds freeline"
-        pass
-    
-    def add_polygon(self):
-        "adds a polygon"
-        pass
+        for feature in self._content:
+            yield feature
+            
+    def stringify(self):
+        "this output is saved to backend"
+
+        return json.dumps(self._content)
+