+++ dokumentaatio/changelog.md
@@ -13,3 +13,6 @@
 - user:pass validation added
 - mimicking todo-app class dynamics + fragmentation of source files
 - drawing dynamics mapped, forget about SVG, using tkinter.Canvas
+- rectangle, oval, line, text (with dummy placeholder) drawing works
+- able to store dwg in JSON (saving rigged on exit)
+- able to load earlier drawing